+++ .replit
@@ -1,1 +1,2 @@
 run = "javac Main.java"
+language = "java18"
