+++ Session/Map/MapModel.java
@@ -2,7 +2,7 @@ package Session.Map;
 
 import javax.swing.JToggleButton;
 import javax.swing.*;
-
+import Session.Map.Chip;
 import Exceptions.InvalidPlayerAmountException;
 /**
  * MVC Pattern in use!
@@ -256,47 +256,5 @@ public class MapModel
       }
     }
     return 0;
-  }
-
-  public class Chip
-  {
-    public static final int MAX_TURN_COUNTER = MAX_PLAYERS;
-    public static final int MIN_TURN_COUNTER = 1;
-    
-    /** *Kkeeps tracks of turns among all chips, a turn represents */
-    private static int _playerTurnCounter = MIN_TURN_COUNTER;
-    /** * Assigned player turn to the Chip Instance */
-    private int _playerTurn = _playerTurnCounter;
-
-    /**
-     * binds the the chip instance to the turn number, then increments the turn counter
-     */
-    public Chip()
-    {
-      _playerTurn = _playerTurnCounter;
-      if(_playerTurnCounter > MAX_TURN_COUNTER) 
-      {
-         _playerTurnCounter = MIN_TURN_COUNTER; 
-      } 
-      else 
-      {
-         _playerTurnCounter++; 
-      }
-    }
-
-    /**
-     * @return a player turn associated with Chip instance (whose player the chip is)
-     */
-    public int getPlayerTurn()
-    {
-      return _playerTurn;
-    }  
-    /**
-     * @return an overall count player turns
-     */
-    public static int getPlayerTurnCounter()
-    {
-      return _playerTurnCounter;
-    }
   }
 }
